Ask for payment method before upgrade
