+++ packages/webapp/src/pages/Team/Billing/components/Payment.tsx
@@ -1,19 +1,14 @@
-import { AddressElement, Elements, PaymentElement, useElements, useStripe } from '@stripe/react-stripe-js';
-import { CreditCard, Loader } from 'lucide-react';
-import { useEffect, useMemo, useState } from 'react';
+import { CreditCard } from 'lucide-react';
+import { useMemo } from 'react';
 
 import { Dot } from './Dot';
+import { PaymentMethodDialog } from './PaymentMethodDialog';
 import { StyledLink } from '@/components-v2/StyledLink';
 import { Button } from '@/components-v2/ui/button';
-import { Dialog, DialogClose, DialogContent, DialogFooter, DialogHeader, DialogTitle, DialogTrigger } from '@/components-v2/ui/dialog';
 import { Skeleton } from '@/components-v2/ui/skeleton';
 import { useApiGetBillingUsage } from '@/hooks/usePlan';
-import { apiPostStripeCollectPayment, useStripePaymentMethods } from '@/hooks/useStripe';
-import { useToast } from '@/hooks/useToast';
-import { queryClient, useStore } from '@/store';
-import { stripePromise } from '@/utils/stripe';
-
-import type { PostStripeCollectPayment } from '@nangohq/types';
+import { useStripePaymentMethods } from '@/hooks/useStripe';
+import { useStore } from '@/store';
 
 export const Payment: React.FC = () => {
     const env = useStore((state) => state.env);
@@ -44,7 +39,11 @@ export const Payment: React.FC = () => {
                             </span>
                         </div>
                     </div>
-                    <PaymentFormDialog replace={!!paymentMethod} />
+                    <PaymentMethodDialog replace={!!paymentMethod}>
+                        <Button size={'sm'} className="min-w-27">
+                            {paymentMethod ? 'Update' : 'Add payment method'}
+                        </Button>
+                    </PaymentMethodDialog>
                 </div>
             )}
 
@@ -58,139 +57,5 @@ export const Payment: React.FC = () => {
                 )
             )}
         </div>
-    );
-};
-
-const PaymentFormDialog: React.FC<{ replace?: boolean }> = ({ replace }) => {
-    const env = useStore((state) => state.env);
-
-    const [open, setOpen] = useState(false);
-    const [clientSecret, setClientSecret] = useState<string | null>(null);
-
-    useEffect(() => {
-        if (open && !clientSecret) {
-            const fetchClientSecret = async () => {
-                const secret = ((await apiPostStripeCollectPayment(env)).json as PostStripeCollectPayment['Success']).data.secret;
-                setClientSecret(secret);
-            };
-            void fetchClientSecret();
-        }
-    }, [open, clientSecret, env]);
-
-    const handleDialogOpenChange = (newOpen: boolean) => {
-        setOpen(newOpen);
-        if (newOpen) {
-            // Reset client secret when opening dialog
-            setClientSecret(null);
-        }
-    };
-
-    return (
-        <Dialog open={open} onOpenChange={handleDialogOpenChange}>
-            <DialogTrigger asChild>
-                <Button size={'sm'} className="min-w-27">
-                    {replace ? 'Update' : 'Add payment method'}
-                </Button>
-            </DialogTrigger>
-            <DialogContent>
-                <DialogHeader>
-                    <DialogTitle>{replace ? 'Update' : 'Add'} payment method</DialogTitle>
-                </DialogHeader>
-                {clientSecret ? (
-                    <Elements
-                        stripe={stripePromise}
-                        options={{
-                            loader: 'always',
-                            appearance: {
-                                labels: 'floating',
-                                variables: {
-                                    colorPrimary: '#00b2e3',
-                                    borderRadius: '4px',
-                                    colorTextPlaceholder: '#8b8c8f',
-                                    colorTextSecondary: '#c4c5c7',
-                                    colorBackground: '#18191b',
-                                    colorText: '#fff',
-                                    focusBoxShadow: 'transparent',
-                                    fontFamily: 'Inter, system-ui, sans-serif',
-                                    fontSizeSm: '12px',
-                                    fontSizeBase: '14px',
-                                    spacingUnit: '4px'
-                                }
-                            },
-                            clientSecret
-                        }}
-                    >
-                        <PaymentForm onSuccess={() => handleDialogOpenChange(false)} />
-                    </Elements>
-                ) : (
-                    <div className="flex flex-col gap-4">
-                        <Skeleton className="w-full h-13 bg-bg-subtle" />
-                        <Skeleton className="w-full h-13 bg-bg-subtle" />
-                        <Skeleton className="w-full h-13 bg-bg-subtle" />
-                        <Skeleton className="w-full h-13 bg-bg-subtle" />
-                        <Skeleton className="w-full h-13 bg-bg-subtle" />
-                        <Skeleton className="w-full h-13 bg-bg-subtle" />
-                    </div>
-                )}
-            </DialogContent>
-        </Dialog>
-    );
-};
-
-const PaymentForm: React.FC<{ onSuccess: () => void }> = ({ onSuccess }) => {
-    const stripe = useStripe();
-    const elements = useElements();
-    const { toast } = useToast();
-    const [loading, setLoading] = useState(false);
-
-    // eslint-disable-next-line @typescript-eslint/no-misused-promises
-    const handleSubmit: React.FormEventHandler<HTMLFormElement> = async (e) => {
-        e.preventDefault();
-
-        setLoading(true);
-
-        if (!stripe || !elements) {
-            toast({ title: 'Stripe not loaded', variant: 'error' });
-            setLoading(false);
-            return;
-        }
-
-        const result = await stripe.confirmSetup({
-            elements,
-            confirmParams: {
-                // No return_url to avoid redirect
-            },
-            redirect: 'if_required'
-        });
-
-        if (result.error) {
-            toast({ title: result.error.message, variant: 'error' });
-        } else {
-            toast({ title: 'Payment method added', variant: 'success' });
-            await queryClient.invalidateQueries({ queryKey: ['stripe'] });
-            onSuccess();
-        }
-
-        setLoading(false);
-    };
-
-    return (
-        <form onSubmit={handleSubmit} className="flex flex-col gap-10">
-            <div className="flex flex-col gap-4 max-h-[70vh] min-h-80 overflow-y-auto overflow-x-hidden flex-1">
-                <PaymentElement />
-                <AddressElement options={{ mode: 'billing' }} />
-            </div>
-            <DialogFooter>
-                <DialogClose asChild>
-                    <Button variant="secondary" size="lg">
-                        Cancel
-                    </Button>
-                </DialogClose>
-                <Button type="submit" disabled={loading} variant={'primary'} size="lg">
-                    {loading && <Loader className="animate-spin" />}
-                    Save payment method
-                </Button>
-            </DialogFooter>
-        </form>
     );
 };
